+++ composer.json
@@ -1,6 +1,6 @@
 {
     "name": "macocci7/php-histogram",
-    "version": "1.2.0",
+    "version": "1.2.1",
     "description": "it's easy to use for creating histograms.",
     "type": "library",
     "license": "MIT",
@@ -19,7 +19,7 @@
     "require": {
         "php": ">=8.1",
         "macocci7/php-frequency-table": "^1.3",
-        "intervention/image": "^3.5",
+        "intervention/image": "^3.6",
         "nette/neon": "^3.4",
         "fakerphp/faker": "^1.23"
     },
@@ -28,6 +28,6 @@
         "phpunit/phpunit": "^10.5",
         "phpmd/phpmd": "^2.15",
         "phpstan/phpstan": "^1.10",
-        "php-parallel-lint/php-parallel-lint": "^1.3"
+        "php-parallel-lint/php-parallel-lint": "^1.4"
     }
 }
