+++ src/Traits/AttributeTrait.php
@@ -1,0 +1,108 @@
+<?php
+
+namespace Macocci7\PhpHistogram\Traits;
+
+trait AttributeTrait
+{
+    protected int $CANVAS_WIDTH_LIMIT_LOWER;
+    protected int $CANVAS_HEIGHT_LIMIT_LOWER;
+
+    protected int $canvasWidth;
+    protected int $canvasHeight;
+    protected float $frameXRatio;
+    protected float $frameYRatio;
+    protected string $labelX;
+    protected string $labelY;
+    protected string $caption;
+
+    /**
+     * returns current canvas size
+     * @return  array<string, int>  [width(px), height(px)]
+     */
+    public function size()
+    {
+        return [
+            'width' => $this->canvasWidth,
+            'height' => $this->canvasHeight,
+        ];
+    }
+
+    /**
+     * resizes the canvas size
+     * @param   int $width  specify in pix at least 50
+     * @param   int $height specify in pix at least 50
+     * @return  self
+     * @thrown  \Exception
+     */
+    public function resize(int $width, int $height)
+    {
+        if ($width < $this->CANVAS_WIDTH_LIMIT_LOWER) {
+            throw new \Exception(
+                "width is below the lower limit "
+                . $this->CANVAS_WIDTH_LIMIT_LOWER
+            );
+        }
+        if ($height < $this->CANVAS_HEIGHT_LIMIT_LOWER) {
+            throw new \Exception(
+                "height is below the lower limit "
+                . $this->CANVAS_HEIGHT_LIMIT_LOWER
+            );
+        }
+        $this->canvasWidth = $width;
+        $this->canvasHeight = $height;
+        return $this;
+    }
+
+    /**
+     * sets the frame ratio of the histogram area to the canvas size
+     * @param   float   $xRatio (0.0 < $xRatio < 1.0)
+     * @param   float   $yRatio (0.0 < $yRatio < 1.0)
+     * @return  self
+     * @thrown  \Exception
+     */
+    public function frame(float $xRatio, float $yRatio)
+    {
+        if ($xRatio <= 0.0 || $xRatio > 1.0) {
+            throw new \Exception("Ratio must be: 0.0 < ratio <= 1.0.");
+        }
+        if ($yRatio <= 0.0 || $yRatio > 1.0) {
+            throw new \Exception("Ratio must be: 0.0 < ratio <= 1.0.");
+        }
+        $this->frameXRatio = $xRatio;
+        $this->frameYRatio = $yRatio;
+        return $this;
+    }
+
+    /**
+     * sets label of X
+     * @param   string  $label
+     * @return  self
+     */
+    public function labelX(string $label)
+    {
+        $this->labelX = $label;
+        return $this;
+    }
+
+    /**
+     * sets label of Y
+     * @param   string  $label
+     * @return  self
+     */
+    public function labelY(string $label)
+    {
+        $this->labelY = $label;
+        return $this;
+    }
+
+    /**
+     * sets caption
+     * @param   string  $caption
+     * @return  self
+     */
+    public function caption(string $caption)
+    {
+        $this->caption = $caption;
+        return $this;
+    }
+}
